Attempt - Add Student function
Changed the outlay of the project:

Now - each student can have an image associated with it

Attempted to create an "Add Student" button, but am struggling with the final implementation of it. I am using the "AddStudentModal" to create a modal background and a modal that will allow us to add new students with default information. Right now, when you click on the "Add Student" button, the modal background will appear, but not the modal itself that allows you to add the student. It took a lot of troubleshooting to even get the background to appear, but after over an hour of research and attempts, I have been unable to get the Modal itself to appear.

The inspect element on my computer shows that the "modal" class is greyed out for some reason that I am struggling to figure out why. Any assistance is welcome and would be greatly appreciated.

-Joe