Created a "Location" element for the students
Created - Location - status:

Location not currently used but can be utilized in future functionality

Notes: Should we have a present vs absent status as well? Or could "Absent" simply be part of the location and we use this to filter absent students to the bottom of the page? Just food for thought